Move cryptography into crypto.py and fix imports
Closes #130.
