+++ privcount/crypto.py
@@ -1,0 +1,308 @@
+'''
+Created on Dec 6, 2016
+
+@author: teor
+'''
+
+import logging
+import datetime
+import uuid
+import json
+
+from math import ceil
+from time import time
+from base64 import b64encode, b64decode
+
+from hashlib import sha256 as DigestHash
+# encryption using SHA256 requires cryptography >= 1.4
+from cryptography.hazmat.primitives.hashes import SHA256 as CryptoHash
+
+from cryptography import x509
+from cryptography.fernet import Fernet
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import serialization, hashes, hmac
+from cryptography.hazmat.primitives.asymmetric import rsa, padding
+from cryptography.exceptions import UnsupportedAlgorithm, InvalidSignature
+
+def load_private_key_string(key_string):
+    return serialization.load_pem_private_key(key_string, password=None, backend=default_backend())
+
+def load_private_key_file(key_file_path):
+    with open(key_file_path, 'rb') as key_file:
+        private_key = load_private_key_string(key_file.read())
+    return private_key
+
+def load_public_key_string(key_string):
+    return serialization.load_pem_public_key(key_string, backend=default_backend())
+
+def load_public_key_file(key_file_path):
+    with open(key_file_path, 'rb') as key_file:
+        public_key = load_public_key_string(key_file.read())
+    return public_key
+
+def get_public_bytes(key_string, is_private_key=True):
+    if is_private_key:
+        private_key = load_private_key_string(key_string)
+        public_key = private_key.public_key()
+    else:
+        public_key = load_public_key_string(key_string)
+    return public_key.public_bytes(encoding=serialization.Encoding.PEM, format=serialization.PublicFormat.SubjectPublicKeyInfo)
+
+def get_public_digest_string(key_string, is_private_key=True):
+    return DigestHash(get_public_bytes(key_string, is_private_key)).hexdigest()
+
+def get_public_digest(key_path, is_private_key=True):
+    with open(key_path, 'rb') as key_file:
+        digest = get_public_digest_string(key_file.read(), is_private_key)
+    return digest
+
+def get_serialized_public_key(key_path, is_private_key=True):
+    with open(key_path, 'rb') as key_file:
+        data = get_public_bytes(key_file.read(), is_private_key)
+    return data
+
+def get_hmac(secret_key, unique_prefix, data):
+    '''
+    Perform a HMAC using the secret key, unique hash prefix, and data.
+    The key must be kept secret.
+    The prefix ensures hash uniqueness.
+    Returns HMAC-SHA256(secret_key, unique_prefix | data) as bytes.
+    '''
+    # If the secret key is shorter than the digest size, security is reduced
+    assert secret_key
+    assert len(secret_key) >= CryptoHash.digest_size
+    h = hmac.HMAC(bytes(secret_key), CryptoHash(), backend=default_backend())
+    h.update(bytes(unique_prefix))
+    h.update(bytes(data))
+    return bytes(h.finalize())
+
+def verify_hmac(expected_result, secret_key, unique_prefix, data):
+    '''
+    Perform a HMAC using the secret key, unique hash prefix, and data, and
+    verify that the result of:
+    HMAC-SHA256(secret_key, unique_prefix | data)
+    matches the bytes in expected_result.
+    The key must be kept secret. The prefix ensures hash uniqueness.
+    Returns True if the signature matches, and False if it does not.
+    '''
+    # If the secret key is shorter than the digest size, security is reduced
+    assert secret_key
+    assert len(secret_key) >= CryptoHash.digest_size
+    h = hmac.HMAC(bytes(secret_key), CryptoHash(), backend=default_backend())
+    h.update(bytes(unique_prefix))
+    h.update(bytes(data))
+    try:
+        h.verify(bytes(expected_result))
+        return True
+    except cryptography.exceptions.InvalidSignature:
+        return False
+
+def b64_raw_length(byte_count):
+    '''
+    Note: base64.b64encode returns b64_padded_length bytes of output.
+    Return the raw base64-encoded length of byte_count bytes.
+    '''
+    if byte_count < 0:
+        raise ValueError("byte_count must be non-negative")
+    return long(ceil(byte_count*8.0/6.0))
+
+B64_PAD_TO_MULTIPLE = 4
+
+def b64_padded_length(byte_count):
+    '''
+    Return the padded base64-encoded length of byte_count bytes, as produced
+    by base64.b64encode.
+    '''
+    raw_length = b64_raw_length(byte_count)
+    # base64 padding rounds up to the nearest multiple of 4
+    trailing_bytes = raw_length % B64_PAD_TO_MULTIPLE
+    if trailing_bytes > 0:
+        padding_bytes = B64_PAD_TO_MULTIPLE - trailing_bytes
+    else:
+        padding_bytes = 0
+    padded_length = raw_length + padding_bytes
+    assert padded_length % B64_PAD_TO_MULTIPLE == 0
+    return padded_length
+
+def encode_data(data_structure):
+    """
+    Encode an arbitrary python data structure in a format that is suitable
+    for encryption (encrypt() expects bytes).
+    The data structure can only contain basic python types, those supported
+    by json.dumps (including longs, but no arbitrary objects).
+    Performs the following transformations, in order:
+    - dump the data structure using json: b64encode doesn't encode objects
+    - b64encode the json: avoid any round-trip string encoding issues
+    Returns a base64 blob that can safely be encrypted, decrypted, then passed
+    to decode_data to produce the original data structure.
+    """
+    json_string = json.dumps(data_structure)
+    return b64encode(json_string)
+
+def decode_data(encoded_string):
+    """
+    Decode an arbitrary python data structure from the format provided by
+    encode_data().
+    The data structure can only contain basic python types, those supported
+    by json.loads (including longs, but no arbitrary objects).
+    Performs the inverse transformations to encode_data().
+    Returns a python data structure.
+    """
+    json_string = b64decode(encoded_string)
+    # json.loads is safe to use on untrusted data (from the network)
+    return json.loads(json_string)
+
+def generate_symmetric_key():
+    """
+    Generate and return a new secret key that can be used for symmetric
+    encryption.
+    """
+    return Fernet.generate_key()
+
+def encrypt_symmetric(secret_key, plaintext):
+    """
+    Encrypt plaintext with the Fernet symmetric key secret_key.
+    This key must be kept secret, as it can be used to decrypt the message.
+    The encrypted message contains its own creation time in plaintext:
+    this time is visible to an attacker.
+    See https://cryptography.io/en/latest/fernet/ for the details of this
+    encryption scheme.
+    Returns the encrypted ciphertext.
+    """
+    f = Fernet(secret_key)
+    return f.encrypt(plaintext)
+
+def decrypt_symmetric(secret_key, ciphertext, ttl=None):
+    """
+    Decrypt ciphertext with the Fernet symmetric key secret_key.
+    See https://cryptography.io/en/latest/fernet/ for the details of this
+    encryption scheme.
+    Returns the decrypted plaintext
+    Throws an exception if secret_key or ciphertext are invalid, or the
+    message is older than ttl seconds.
+    """
+    f = Fernet(secret_key)
+    # fernet requires the ciphertext to be bytes, it will raise an exception
+    # if it is a string
+    return f.decrypt(bytes(ciphertext), ttl)
+
+def encrypt_pk(pub_key, plaintext):
+    """
+    Encrypt plaintext with the RSA public key pub_key, using CryptoHash()
+    as the OAEP/MGF1 padding hash.
+    plaintext is limited to the size of the RSA key, minus the padding, or a
+    few hundred bytes.
+    Returns a b64encoded ciphertext string.
+    Encryption failures result in an exception being raised.
+    """
+    try:
+        ciphertext = pub_key.encrypt(
+            plaintext,
+            padding.OAEP(
+                mgf=padding.MGF1(algorithm=CryptoHash()),
+                algorithm=CryptoHash(),
+                label=None
+                )
+            )
+    except UnsupportedAlgorithm as e:
+        # a failure to encrypt our own data is a fatal error
+        # the most likely cause of this error is an old cryptography library
+        # although some newer binary cryptography libraries are linked with
+        # old OpenSSL versions, to fix, check 'openssl version' >= 1.0.2, then:
+        # pip install -I --no-binary cryptography cryptography
+        logging.error("Fatal error: encryption hash {} unsupported, try upgrading to cryptography >= 1.4 compiled with OpenSSL >= 1.0.2. Exception: {}".format(
+                          CryptoHash, e))
+        # re-raise the exception for the caller to handle
+        raise e
+    return b64encode(ciphertext)
+
+def decrypt_pk(priv_key, ciphertext):
+    """
+    Decrypt a b64encoded ciphertext string with the RSA private key priv_key,
+    using CryptoHash() as the OAEP/MGF1 padding hash.
+    Returns the plaintext.
+    Fails and calls os._exit on an UnsupportedAlgorithm exception.
+    (Other decryption failures result in an exception being raised.)
+    """
+    try:
+        plaintext = priv_key.decrypt(
+            b64decode(ciphertext),
+            padding.OAEP(
+                mgf=padding.MGF1(algorithm=CryptoHash()),
+                algorithm=CryptoHash(),
+                label=None
+                )
+            )
+    except UnsupportedAlgorithm as e:
+        # a failure to dencrypt someone else's data is not typically a fatal
+        # error, but in this particular case, the most likely cause of this
+        # error is an old cryptography library
+        logging.error("Fatal error: encryption hash {} unsupported, try upgrading to cryptography >= 1.4. Exception: {}".format(
+                          CryptoHash, e))
+        # re-raise the exception for the caller to handle
+        raise e
+    return plaintext
+
+def encrypt(pub_key, data_structure):
+    """
+    Encrypt an arbitrary python data structure, using the following scheme:
+    - transform the data structure into a b64encoded json string
+    - encrypt the string with a single-use symmetric encryption key
+    - encrypt the single-use key using asymmetric encryption with pub_key
+    The data structure can contain any number of nested dicts, lists, strings,
+    doubles, ints, and longs.
+    Returns a data structure containing ciphertexts, which should be treated
+    as opaque.
+    Encryption failures result in an exception being raised.
+    """
+    encoded_string = encode_data(data_structure)
+    # TODO: secure delete
+    secret_key = generate_symmetric_key()
+    sym_encrypted_data = encrypt_symmetric(secret_key, encoded_string)
+    pk_encrypted_secret_key = encrypt_pk(pub_key, secret_key)
+    return { 'pk_encrypted_secret_key': pk_encrypted_secret_key,
+             'sym_encrypted_data': sym_encrypted_data}
+
+def decrypt(priv_key, ciphertext):
+    """
+    Decrypt ciphertext, yielding an arbitrary python data structure, using the
+    same scheme as encrypt().
+    ciphertext is a data structure produced by encrypt(), and should be
+    treated as opaque.
+    Returns a python data structure.
+    Decryption failures result in an exception being raised.
+    """
+    pk_encrypted_secret_key = ciphertext['pk_encrypted_secret_key']
+    sym_encrypted_data = ciphertext['sym_encrypted_data']
+    # TODO: secure delete
+    secret_key = decrypt_pk(priv_key, pk_encrypted_secret_key)
+    encoded_string = decrypt_symmetric(secret_key, sym_encrypted_data)
+    return decode_data(encoded_string)
+
+def generate_keypair(key_out_path):
+    private_key = rsa.generate_private_key(public_exponent=65537, key_size=4096, backend=default_backend())
+    pem = private_key.private_bytes(encoding=serialization.Encoding.PEM, format=serialization.PrivateFormat.PKCS8, encryption_algorithm=serialization.NoEncryption())
+    with open(key_out_path, 'wb') as outf:
+        print >>outf, pem
+
+def generate_cert(key_path, cert_out_path):
+    private_key = load_private_key_file(key_path)
+    public_key = private_key.public_key()
+
+    builder = x509.CertificateBuilder()
+    builder = builder.subject_name(x509.Name([
+        x509.NameAttribute(x509.OID_COMMON_NAME, u'PrivCount User'),
+    ]))
+    builder = builder.issuer_name(x509.Name([
+        x509.NameAttribute(x509.OID_COMMON_NAME, u'PrivCount Authority'),
+    ]))
+    builder = builder.not_valid_before(datetime.datetime.today() - datetime.timedelta(days=1))
+    builder = builder.not_valid_after(datetime.datetime(2020, 1, 1))
+    builder = builder.serial_number(int(uuid.uuid4()))
+    builder = builder.public_key(public_key)
+    builder = builder.add_extension(x509.BasicConstraints(ca=False, path_length=None), critical=True)
+
+    certificate = builder.sign(private_key=private_key, algorithm=hashes.SHA256(), backend=default_backend())
+
+    with open(cert_out_path, 'wb') as outf:
+        print >>outf, certificate.public_bytes(encoding=serialization.Encoding.PEM)
