+++ privcount/data_collector.py
@@ -13,10 +13,11 @@ from twisted.internet import task, reactor, ssl
 from twisted.internet.protocol import ReconnectingClientFactory
 
 from privcount.counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
+from privcount.crypto import get_public_digest_string, load_public_key_string, encrypt
 from privcount.log import log_error, format_delay_time_wait, format_last_event_time_since
 from privcount.node import PrivCountClient
 from privcount.protocol import PrivCountClientProtocol, TorControlClientProtocol
-from privcount.util import get_public_digest_string, load_public_key_string, encrypt, normalise_path, choose_secret_handshake_path
+from privcount.util import normalise_path, choose_secret_handshake_path
 
 # using reactor: pylint: disable=E1101
 # method docstring missing: pylint: disable=C0111
