+++ test/test_counter.py
@@ -3,11 +3,11 @@
 # this test will fail if any counter inconsistencies are detected
 # 
 
+import sys
+
 from privcount.counter import SecureCounters, adjust_count_signed, counter_modulus, add_counter_limits_to_config, get_events_for_known_counters
 from math import sqrt
 from random import SystemRandom
-
-import sys
 
 import logging
 # DEBUG logs every check: use it on failure
