+++ test/test_encryption.py
@@ -5,14 +5,14 @@
 # this test will exit successfully if the decrypted data matches the original
 # encrypted data
 
+import string
+import sys
+
 from base64 import b64encode, b64decode
 from os import urandom
 from random import SystemRandom
 
-import string
-import sys
-
-from privcount.util import load_public_key_file, load_private_key_file, encrypt_pk, decrypt_pk, generate_symmetric_key, encrypt_symmetric, decrypt_symmetric, encode_data, decode_data, encrypt, decrypt
+from privcount.crypto import load_public_key_file, load_private_key_file, encrypt_pk, decrypt_pk, generate_symmetric_key, encrypt_symmetric, decrypt_symmetric, encode_data, decode_data, encrypt, decrypt
 from privcount.counter import counter_modulus
 
 import logging
